fix: saveState on private key login
